requirements.txt
psycopg2-binary>=2.9.3
python-dotenv>=0.19.0
Faker>=8.0.0
